Add NewFromGoFuzz
Add a helper function that enables using gofuzz (this project) with
go-fuzz (https://github.com/dvyukov/go-fuzz) for continuose
fuzzing. Essentially, it enables translating the fuzzing bytes from
go-fuzz to any Go object using this library.

This change will enable using this project with fuzzing websites such as fuzzit.dev
or fuzzbuzz.io.

The underlying implementation was an idea of lavalamp, by which a random source
is created that generates random numbers from the input bytes slice. In this way
changes in the source result in logical changes in the random data, which enables
the fuzzer to efficiently search the space of inputs.

Fixes #33