Fix charRange to allow picking the last item (#36)
* Add test proving charRange.choose excludes the last item

Change charRange#choose to accept an interface, int63nPicker, exposing only
Int63n(int64) int64. This allows the test to pass a custom RNG, returning a
constant value (0 or the last possible int).

Change charRange#choose to act on an immutable copy of charRange.

* Allow charRange.choose to pick the last rune

Before this change, choose would never pick the last character of the set. This
is not the expected behavior: if the range is a-z, we expect a,b,...,y,z to be
picked in a uniform way.