Allow Continue.Fuzz to accept a reflect.Value
Continue.Fuzz (and FuzzNoCustom) immediately convert the interface{}
into a reflect.Value. In some cases the caller may already have a
reflect.Value, so accepting that type makes it easier to write custom
fuzz functions.

One use case for this is ignoring a field on a struct. The SkipFieldsWithPattern
option is a good choice when the field name should always be ignored. In other
cases a field name may exist on multiple structs, and using a regex pattern
would result in the field being skipped in all cases. To ignore a field by name
on only a single struct a custom fuzz function can use reflect to iterate over
the fields, call c.Fuzz() on all fields except for the ignored field.