Reduce the chance a test can flake
This commit adjust the values used in tests to reduce the probability of a flake.

For `tryFuzz`, increase the attempts to 30.
For `checkFailed`, increase the attempts to 20, and only fail if 19/20 iteration received a zero value. Previously it would fail if 9/10 received zero values.

Also adds t.Helper() to the two test helpers.